third_party/blink/web_tests/platform/generic/fast/forms/number/number-spinbutton-change-and-input-events-expected.txt

Test for event dispatching by spin buttons in a type=number input.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Initial state
PASS changeEventCounter is 0
PASS inputEventCounter is 0
Click the upper button
Triggers only input event on mouseDown
PASS testInput.value is "1"
PASS changeEventCounter is 0
PASS inputEventCounter is 1
Triggers only change event on mouseUp
PASS testInput.value is "1"
PASS changeEventCounter is 1
PASS inputEventCounter is 1
Click again, but the value is not changed.
PASS testInput.value is "1"
PASS changeEventCounter is 1
PASS inputEventCounter is 1
Focus on another field
PASS changeEventCounter is 1
PASS inputEventCounter is 1
PASS successfullyParsed is true

TEST COMPLETE
